requirements.txt
alembic
jupyterlab
numpy
pandas
psycopg2
python-dotenv
requests
SQLAlchemy
